Updated database support
Servers and users will now correctly be added to the database when needed.

Servers and users tables will be refilled everytime the bot is started to avoid storing unused data in the database

Questions table will not be refilled so that data is preserved even if bot is kicked and reinvited from server.